refactor eliminate element.tsx
