+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListContent/Row/styles.module.scss
@@ -1,0 +1,16 @@
+@import "$styles/_variables";
+@import "$styles/mixins/_lists";
+
+.row {
+  @include applicantListRows;
+
+  padding: $spacing-s;
+
+  &.grey {
+    background-color: $lighterGrey;
+  }
+
+  &:hover {
+    background-color: $selectedColor;
+  }
+}
