+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListContent/component.tsx
@@ -1,23 +1,12 @@
 import React, { FunctionComponent } from "react";
-import { Columns } from "../Columns";
-import { Row, Item } from "./elements";
+import { Row } from "./Row";
+import { Item } from "./Item";
 import { IApplicant } from "$interfaces/Applicant";
 
 export const ListContent: FunctionComponent<IListContent> = ({ applicants }) => (
   <Row applicants={applicants}>
     {
-      applicant => (
-        <Columns key={applicant.uuid}>
-          {
-            column =>
-              <Item
-                key={`${applicant.uuid}-${column}`}
-                column={column}
-                applicant={applicant}
-              />
-          }
-        </Columns>
-      )
+      applicant => <Item applicant={applicant}/>
     }
   </Row>
 );
