+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListContent/container.tsx
@@ -1,4 +1,4 @@
-import React, { FunctionComponent, Fragment } from "react";
+import React, { FunctionComponent } from "react";
 import { useGetApplicants } from "$hooks";
 import { ListContent } from "./component";
 
@@ -6,8 +6,8 @@ export const ListContentContainer: FunctionComponent = () => {
   const result = useGetApplicants();
   const applicants = result?.data?.getApplicants;
   return (
-    <Fragment>
+    <>
       {applicants && <ListContent applicants={applicants} />}
-    </Fragment>
+    </>
   );
 };
