+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListHeader/HeaderContainer/index.tsx
@@ -1,0 +1,8 @@
+import React, { FunctionComponent } from "react";
+import styles from "./styles.module.scss";
+
+export const HeaderContainer: FunctionComponent = ({ children }) => (
+  <div className={styles.headerContainer}>
+    {children}
+  </div>
+);
