+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListHeader/HeaderContainer/styles.module.scss
@@ -1,0 +1,12 @@
+@import "$styles/_variables";
+@import "$styles/mixins/_lists";
+
+.headerContainer {
+  @include applicantListRows;
+
+  background-color: $white;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.25);
+  padding: $spacing-s;
+  position: sticky;
+  top: 0;
+}
