+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListHeader/Item/index.tsx
@@ -2,12 +2,6 @@ import React, { FunctionComponent } from "react";
 import classNames from "classnames";
 import styles from "./styles.module.scss";
 
-export const HeaderContainer: FunctionComponent = ({ children }) => (
-  <div className={styles.headerContainer}>
-    {children}
-  </div>
-);
-
 interface IItem {
   column: string;
   text: string;
