+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListHeader/component.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from "react";
-import { HeaderContainer, Item } from "./elements";
+import { HeaderContainer } from "./HeaderContainer";
+import { Item } from "./Item";
 import { ITranslations } from "./container";
 import { Columns } from "../Columns";
