+++ src/pages/Admin/ApplicantSearch/ApplicantList/ListHeader/container.tsx
@@ -1,13 +1,13 @@
-import React, { FunctionComponent, Fragment } from "react";
+import React, { FunctionComponent } from "react";
 import { useTranslations } from "$hooks";
 import { ListHeader } from "./component";
 
 export const ListHeaderContainer: FunctionComponent = () => {
   const translations = useTranslations<ITranslations>("adminApplicantListHeader");
   return (
-    <Fragment>
+    <>
       {translations && <ListHeader translations={translations} />}
-    </Fragment>
+    </>
   );
 };
