+++ src/pages/Admin/ApplicantSearch/ApplicantList/component.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react";
 import { ListHeader } from "./ListHeader";
 import { ListContent } from "./ListContent";
-import { ListContainer } from "./elements";
+import { ListContainer } from "./ListContainer";
 
 export const ApplicantList: FunctionComponent = () => (
   <ListContainer>
