+++ src/pages/Admin/ApplicantSearch/MainContainer/styles.module.scss
