+++ src/pages/Admin/ApplicantSearch/component.tsx
@@ -1,20 +1,17 @@
 import React, { FunctionComponent } from "react";
 import { Window } from "$components/Window";
 import { MainTitle } from "./MainTitle";
-import { MainContainer } from "./elements";
+import { MainContainer } from "./MainContainer";
 import { ApplicantList } from "./ApplicantList";
 
-export const ApplicantSearch: FunctionComponent<IApplicantSearch> = () => {
-
-  return (
-    <Window width="fullWidth" desktopOnly>
-      <MainContainer>
-        <MainTitle />
-        <ApplicantList />
-      </MainContainer>
-    </Window>
-  );
-};
+export const ApplicantSearch: FunctionComponent<IApplicantSearch> = () => (
+  <Window width="fullWidth" desktopOnly>
+    <MainContainer>
+      <MainTitle />
+      <ApplicantList />
+    </MainContainer>
+  </Window>
+);
 
 interface IApplicantSearch {
   translations: {
