--- src/pages/Admin/ApplicantSearch/elements.tsx
@@ -1,12 +1,0 @@
-import React, { FunctionComponent, ReactNode } from "react";
-import styles from "./styles.module.scss";
-
-interface IMainContainer {
-  children?: ReactNode;
-}
-
-export const MainContainer: FunctionComponent<IMainContainer> = ({ children }) => (
-  <div className={styles.mainContent}>
-    {children}
-  </div>
-);
